Hack to trace do_exit (not a syscall); support for dumping a guest ps list.
First: the hack is still a little broken.  If a do_exit is traced last,
then the prober is killed, for whatever reason the guest kernel panics.
The only reason for such a panic should be because the breakpoint
instruction doesn't get replaced correctly.  The only reason I can
figure for that is that perhaps the kernel page table isn't being used
to replace in xenaccess.  If, for instance, one or more do_exits are
traced, then a sys_waitpid or two, there is no problem with killing the
probe.  So probably some assumption I make in tracing syscalls (i.e.,
coming from userspace to kernel space) is breaking...