Correctly count number of pages requested so we can munmap correctly.
Also, when we are looking for NULL-terminated byte strings and don't
know how long the string will/might be, don't just try to map a whole
page.  Why?  We can expect that the page containing the virtual address
in the memory access request will be in the user page table, BUT the
following page may not.  So, initially, only map in that first page,
don't just try for PAGE_SIZE - 1 (since that will force us to ask for
two pages immediately if we need more than two bytes).

And some more debug stuff...