New features (batch probe reg/unreg and sig handling) and a bugfix.
Added two new functions, register_vmprobe_batch and
unregister_vmprobe_batch that register/unregister lists of probes
for a single domain.  The upshot is that the domain only is paused
once, and so probe insertion/deletion is transactional.

When batch registering, there are multiple modes to cover the case
where some probes have been successfully inserted, but then we fail
to insert another.  The caller can specify to ignore the failed
registration and continue; have registration stop immediately but
leave the successfully-registered probes in; or roll back all
successfully-registered probes.

Also, I had to fix up signal handling.  Basically, before, if we
took an interrupt while handling a debugger event, things could
get screwed up in the guest, because we wouldn't allow handlers
to finish before we unregistered probes.  So for now, the signal
handler sets global interrupt vars, but does not unregister probes;
it waits for run_vmprobes to notice the interrupt and lets it do
the registration.  That's not quite right either, but it matches
the current usage model much better.

The bug fix make sure we're not improperly resetting the IP to
the "original" (saved) value when removing a probe.