Add symbol & location lookup/resolve/load; raw values; target register ops and metadata.
The bulk of lookup work is done by the internal function
__symbol_get_one_member, which handles complexities in member lookup
like unnamed struct/union members.  Most other lookup functions call
this one eventually.  Since sometimes for lookups of members, such as
var.x (where var is a struct, and there are two unnamed structs leading
to x), we actually need the unnamed struct variable symbols to resolve
the location of the member --- we add a new data structure called
symbol_chain that is basically a list of nested symbols.  It is tempting
to think that we could just build up an actual tree where the members
point to parent symbols, but we can't, because the symbol for a member
variable is really a member of the *type* symbol for the parent variable
that contains it -- and this type may not only be used for the parent.
So we have to have a separate "chain" for each variable symbol that has
a parent hierarchy.  symbol_chains may have more applications in the
future; i.e., nested functions.  It is currently exposed to the library
user, but I'll mask this via a better data structure later so it is no
longer exposed.