Add support for location lists (and partial support for ranges).
Location lists are used to specify the locations of symbols in a
PC-relative manner (i.e., the value of a local var might be in a
register at some point during a function, but on the stack at
another point in the same function).

A range can either be a simple (lowpc,highpc) tuple, or it can be
a list of begin and end addresses.  This commit just adds the data
structures and some other stuff; we're not actually loading the
.debug_ranges section yet so that we can retrieve the lists.