Intelligent debuginfo source search.
Open up the actual ELF binary and look for three sections to inform
our search.  First, if there is a nonzero .debug_info section,
load that.  Second, if there is a .note.gnu.build-id section,
read the build id and decompose it into a one-byte-dir/file.debug
string that we look for in our search path (i.e., we look for
$PATH/.build-id/b1/b2..bX.debug).  Otherwise, if there is a
.gnu_debuglink section, we read that section and try to find a
matching debug file in $PATH.  The default path is
/usr/lib/debug:/usr/local/lib/debug .