Add support for enabling/disabling use of ELF strtab (DWDEBUG_USE_STRTAB).
Many of the strings we need (i.e., symbol names) are already in the ELF
strtab section.  Thus, copying the whole strtab in one shot reduces
the potentially thousands (or millions) of mallocs we might have to do
to save off each symbol's name.

However, in DWARF, a structure type's name is just the 'x' part of
'struct x'.  This means that even if 'x' was an index into the strtab
and we could save a pointer to it instead of creating our own copy,
we could not insert it into our symtabs as plain 'x' -- typedefs
screw that up potentially (i.e., 'typedef struct x x').  So, our
symtab key has to be 'struct x', so we have to create an "extended"
symbol name.  Thus, this commit adds an extra field in the symbol
struct (but only in the type subsection of the main union, so it only
costs for types) that allows us to do extended names.

Thus, if an extended name is set, symbol.name is just a pointer into
that string, and symbol.s.ti.extname has the extended name.

I mostly added this define to study the tradeoff between saving off
the entire strtab (even though we might not use some of it, and instead
have to malloc extended names for some types), vs not saving the
strtab at all and always malloc'ing.  Looks like saving the strtab is
probably going to be the wise choice...