Handle cases where the debuginfo includes duplicate symbols.
For now we just punt, and stick the duplicate symbols in the anontab
instead of the primary tab for the symtab we're operating on.

Later, when we actually reuse type symbols across CUs, this can be
relaxed, but until there is a design for partial symbol table loading
and unloading, we need to just do this.