Bugfixes for LLVM dwarf: "indirect" function symtabs; fb value direct in reg.
A basic program compiled with LLVM now can be handled.  When emitting
DWARF for a function, LLVM at least sometimes spits out an immediate
lexical_scope tag, and in my symtab hierarchy, this results in a subtab of
the function's primary symtab.  We weren't searching subtabs for local
function vars, but we now search one level IF the subtab is anonymous
(otherwise its name should be included in the search string).

Also, we weren't handling the case where the frame base address value is
contained directly in a register (the one DWARF location op that doesn't
require a resolve/load operation pair).