Add memranges; each memregion can have multiple memranges.
I realized when I was looking at my bounds checking and relocation
code that I had missed one kind of data structure I needed.

Short story: when we lookup symbols on targets, we find them in
debugfile structures, which we have attached to regions.  Before
this commit, a region was basically an address range with protection
flags that was bound to a debugfile... so once we had looked up the
symbol, we knew which region it was attached to.  Except we really
didn't, since debugfiles *must* attach to multiple regions (i.e.,
in the case of loading an ELF object whose segments get mapped into
different ranges of virtual memory).  So we were just blindly
selecting the first region that was associated with the debugfile
containing the symbol... when the symbol might actually have been
in a different region, but in the same debugfile.

So, now regions have become ranges, and regions now contain multiple
ranges.  Regions are now the link between targets and debugfiles.
Thus, when we lookup symbols on a target, until we resolve their
locations, we only know they are in some region -- not which *range*
in that region they are in.  Once we resolve their location, we
can figure out which range they are in.

A lot of this is primarily (only?) useful for shared libraries whose
segments can be relocated, sigh.  But now it's correct.

The additional level of hierarchy in the model doesn't hurt anything,
and it gives us the appropriate level of flexibility for both
objects that might have relocated, non-contiguous segments in memory,
as well as simple, contiguous fixed-address segments.