Handle region/range shifts during location resolution/load.
Obviously, when loading symbol values, we know the initial location
of the symbol's value in memory (and thus its addrspace/region/range), but
if the symbol was a pointer (and if we are auto deref'ing ptrs), an argument,
a member (etc?), we may resolve a location to a new address that is *not*
in the region the caller told us the initial symbol's location was.  This
commit makes sure we handle such transitions.