Add support for HW/SW function breakpoints, and dump args.
Well, we dump the arg location as best we can, BUT for args that
are in registers that get moved onto the stack in the prologue, we
are dumping stack locations that haven't yet been written (see my
Feb 4 emails).

So, it's nice but kinda useless right now, except for breakpoint
testing.