Add location_resolve_loclist to make sure we can load values in registers.
This only handles figuring out which location on a loclist is valid,
given the current EIP.  This is useful in case we are trying to
resolve a location that is a loclist, and the matching location in the
list is a register (which has no address, so location_resolve fails!).
So, this function works in tandem with location_resolve() in
location_load().