Support for monitoring multiple domains.
Config file now supports a "Domain <name>" line which says that all the
following rules, up to the next Domain line, apply to this domain. Support
required moving a bunch of global code into a new per-domain domain_info
struct. Functions to watch, Filters, and Processes (for the process list),
and schandles are now per-domain. Note that the symbol table is still global;
you can only watch multiple domains running the same code, specifically the
domU kernel.

Added a "cookie" to the vmprobes interfaces so we can attach higher-level
state with the vmprobes. Right now there is an explicit get_/set_cookie
interface, but set could be folded into the probe register function and
the cookie should be passed to the handler as a param.

Robustness stuff: if the config file has no probes don't call run_vmprobes,
instead do a sigsuspend waiting for them to indicate that they have changed
the config file. Handle the failed configfile case a little better.

Replace "signal" usage with sigaction and make sure it is properly maintained
across config file signals.

There are still serious robustness problems! Expect that somewhere along the
line you are going to lose a domain or two when you SIGUSR2 vmprobegeneric
to reread the config. Workin on it...