Add support for parsing line number info to obtain prologue and epilogue info.
The main parsing function in dwarf_elfutils.c came from readelf, but I
improved and extended it to deal with prologue/epilogue stuff.

For now, since relatively recent gccs don't generate prologue_end and
epilogue_begin line number program operations, we "assume" which line
number program address corresponds to the end of the prologue.  In practice,
the second address gcc generates in the line number program immediately
after generating a function's lowest address seems to (always or usually) be
the end of the prologue.  So we just assume that it is.

However, I did add support for properly handling prologue_end and
epilogue_begin; LLVM generates the former, and we use that and it works.
If prologue_begin operations exists in the line number program, we use those
in preference to "auto" detection above.