Add support for prologue disassembly; make return actions (aborts) work.
Here's the history behind this commit.  Since we now have to execute
function prologues to make sure we can read the arguments (since arg
locations sometimes haven't yet been populated since the prologue has not
been run), return actions are no longer trivial.  We may be trying to
return from *within* the function body, so the stack pointer has
been advanced by the prologue.  We have to fix this before we
preemptively return; otherwise the stack will get smashed.

Now the logic.  If we don't have a prologue, do the same old immediate
return.  If we do have a prologue, and it starts with a frame pointer
push, assume that we're using frame pointers (i.e., no -fomit-frame-pointer)
and that we can simply set ESP to EBP before single-stepping the ret
instruction.  If the prologue does not start with this, assume that we
have to disassemble the prologue to figure out how the stack pointer
changes.

Our disassembly currently only supports offset changes to the stack
pointer.  We do not run the disassembly each time the action is taken
(which would be necessary to actually emulate the prologue so we could
definitely track all changes to the stack pointer), so we can't handle
instructions like "mul 0x5, %esp".  Fortunately, these don't show up in
prologues.

We try to be really good about catching any arithmetic instructions like
the above whose destination is ESP.

This commit also includes fixes and enhancements to the action code; I
think the biggest fix is to pull in Mike's comment and loop-continue fix
in handle_actions.