Fix several bugs from the last commit; now it's really working.
Don't resume after additional single steps when handling actions; single
step does that itself now.

Major doofus bug: if we're using a frame pointer, we have to call
LEAVE too, not just RET.  It's either that or read the stack ourself; we
should do this anyway (probably) to save single stepping.

In the case of undoing prologue ESP changes, read ESP and update it;
don't read EBP and update that into ESP.

When updating ESP, negate the offset first, before adding it to
current ESP.  Meant to do it and forgot.

Actually enable the "prologue used" bit so we exercise all this code!