Lots of work on trying to get Xen watchpoints working.
Some of these are correctness fixes; fixed up handling of debug registers
like I did for the linux userspace target last week.

But still, Xen watchpoints don't work.  HW breakpoints work fine, but that is
because we fall back to checking the contents of the debug registers for a
match against current EIP.  Obviously, that doesn't work for watchpoints,
since EIP has nothing to do with memory address.  So we have to rely on the
debug status register to tell us which debug register's address triggered the
exception.

But, Xen seems to be wiping out the bits that should be set in the debug
status register.

What's worse is that removing a HW watchpoint will cause the domain to crash
(somehow its EIP gets set to 0x1).  This doesn't happen with HW breakpoints,
and the code path to remove those vs HW watchpoints is identical, as far as
mucking with the debug registers.

Time to abandon this for now, until I or somebody has time to dig into what
is happening to the debug status register.