For now, always force single step even for HW breakpoints.
If we try to skip single step mode for hardware breakpoints IF there are no
post handlers specified for the probepoint, the Xen VM backend will get
stuck in an infinite loop at the breakpoint location.

For whatever reason, this does not happen for the Linux userspace process
target.  Until we know why, just always single step here.