Actually write linux_userproc_launch to launch and attach to the debugged process.
_launch() was empty for a long time; now it works.  So, you can directly launch
binaries instead of quick-attaching to them.

The interesting thing (which may screw up some binaries) is that we try hard to
make sure program loading/dynamic linking has been completed BEFORE we instrument
the program and give control to the user.

On Linux, I can infer this by watching the syscall sequences depending on if
the executable is dynamically or statically linked; I hope this simple scheme
will work.

The alternative (and what has to happen anyway, generally, for targets), is that
every time the target's memory changes (i.e., for userspace, brk, mmap, munmap,
etc; for xen, something else that tells us that the page set has changed), we have
to re-scan the target, adding and removing debugfile sources, expanding and
contracting known address ranges, etc.  It's kind of like swapmod in Emulab :)...
and it's not going to be easy.  So we'll punt it as long as possible.