Change symbol lookup to prefer definitions over declarations.
Usually, when somebody wants to lookup a symbol, they want its location...
not one of the many declaration vars (i.e., a global referenced as
extern in some file, but defined elsewhere).  We weren't distinguishing.

So now, search works like this: if we find a type, return immediately.
If we find a non-type declaration, continue searching through all our
children looking for a definition.  If we don't find one, return a
declaration if we found one.

This has the potential to be slow in a few corner cases, I suppose, but
for the primary use case (extern globals), it will be as fast as usual.

Also, this commit creates rfilters, which are lists of regular expressions
that create a first-match filter.  They work like this:

Each rfilter has a default state where if none of the regular expressions
in its list match.  If the default state is 1 and no list items match, we
accept.  If the default state is 0 and no filter entries match, we reject.
If a list item regex matches, we do whatever the individual filter says.

This is useful in a lot of the kind of searches we need.