Major improvements to the probe API.
This commit includes a fast implementation of the hierarchical probes
idea we discussed last week.  Users can register probes on either
symbols or addresses, which implies the creation of a probepoint to
handle the low-level target manipulation.  Users can also register
probes atop other probes, so you can build up probe forests.

Probes can now have lifecycles.  When a probe is created, the user can
specify a partially-filled 'probe_ops' struct.  The operations in that
table will be invoked at the proper times during the probe's lifetime:
init, registered, enabled, disabled, unregistered, fini (see the
comments in probe_api.h for more detail).  If the ops table is
underspecified or is NULL, nothing happens -- there are no "default"
operations.

Handlers now are invoked with their probe, their probe-specific data (a
void *), and the actual probe that triggered them (if they are just
associated with a probepoint, the trigger probe is themself; if they are
registered on one or more other probes, the trigger probe is the one
they are registered on that was triggered by some propepoint beneath it
in the probe forest.

In the past, probes were created as a side effect of registering some
handlers at a specific address.  This is no longer the case.  Now, the
user must specifically create a probe, with handlers, probe-specific
data, and a probe_ops operations table.  Probes can also be `autofree'
probes, which means that they will be automatically freed by the library
when either they have no more consumers listening to them, or if there
is an error registering them, etc.  There are more comments explaining
this.  It is a complex feature, and might go away... but properly
removing a probe forest in the event of a SEGV is difficult, and I hope
this will make it easier.  The user just has to follow the rules about
holding references to probes.

The current workflow is something like this:

probe =
probe_create(target,probe_ops,"myprobe",pre_handler,post_handler,autofree);
function_bsymbol =
target_lookup_sym(target,symname,delim_char,NULL,SYMBOL_TYPE_FUNCTION);
probe_register_symbol(probe,function_bsymbol,PROBEPOINT_FASTEST,PROBEPOINT_EXEC,PROBEPOINT_LAUTO);

while (1) {
    target_monitor(target);
}

probe_unregister(probe);
if (!autofree)
    probe_free(probe); /* user can just free; that will unregister */