Apply similar fix from lup to make sure our context cache is invalidated.
If the context hadn't been flushed because it was not dirty, but it was
marked as valid, we would have just kept using it after the next target
exception.

Ah, I understand now... the only time when this could have happened is if
we couldn't find any breakpoints or watchpoints to handle, so we bumped
the exception to the user.  If they didn't change our context before
calling target_resume, then the context would have stayed valid into the
next exception, and we would have been looking at old register context.