[CTXPROBES] Provide Parent Task Chain to User Probe Handler
User probe handler receives the current task info through the 'task'
argument.  The user can access to the information of the current
task's parent through 'task->parent'.  If the task does not have
its parent, for example it is the swap task, then 'task->parent' is
NULL.  Take the following code as an example of printing out the chain
of parent tasks:

    while (task->parent)
    {
	    printf("%d (%s)\n", task->parent->pid, task->parent->comm);
		task = task->parent;
	}