Build line number lookup structs and add lookup functions.
Our line number search returns the closest previous line number match
(inclusive).  So, lookup(125) = 120 if that is the closest previous
match; or lookup(125) = 125, etc.

The new lookup functions are debugfile_lookup_addrs_line, which returns
an array_list of object addrs as the data pointers -- since a line may
have multiple addrs associated with it; and debugfile_lookup_sym_line,
which returns the first symbol that matches an addr associated with the
looked-up line.

Also, fix a bug in debugfile_lookup_addr.