Add some new member lookup functions, and a notion of dynamic symbols.
The new functions are

  lsymbol_lookup_member(lsymbol,name,delim)
  symbol_lookup_member(symbol,name,delim)

Dynamic symbols are symbols that our libraries create -- i.e., if you
have a struct type symbol, and you have an address that holds a pointer
to a struct of that type (but it is a dynamically-allocated struct, so
is not necessarily bound to a variable), you are going to want to load
that address as a value of the struct type -- but to do that, you have
to create a dynamic type that points to the struct type you want.  The
target library has a function that builds such symbols -- and it is
unforunately target-specific because we need the target's pointer word
size to build the pointer appropriately.

Also, add some const qualifiers as appropriate.

Also, handle NULL delimiters in lookup functions by using
DWDEBUG_DEF_DELIM (".") instead.