CFI improvement and linux pid dumping in dumptarget.
In dumptarget, for CFI analysis, don't instrument return sites of root
functions we are tracing.  If we instrument these, we will see failed
attempts to pop the shadow stack to check return addr, because the calls
to those functions have (likely) not been instrumented unless there is a
call loop involving one of those root functions and a function it calls.

Also, print out handler_data value in probe handlers for testing
purposes.

Also, for Xen targets, print out the the current task's pid in probe
pre-handler for simple function entry point tracing (the default probing
that happens if you just specify a plain function name on the command
line).  This uses all the new stuff I just committed.