Extend parent process search to real_parent if -r is passed on command line.
If a process in a guest is being ptraced, the kernel sets its
task_struct->parent pointer to the ptracing process, and sets
task_struct->real_parent to the actual parent.  We were only searching
the hierarchy via ->parent; extend the search to use ->real_parent if -r
is specified on the command line.