Implement a quick single step stack; solves the phantom bug for now.
If we hit a breakpoint when we think we should be handling a single step
that we caused to happen, an interrupt (or something else) probably
nailed us before the single step completed.  If that interrupt changes
control flow (which it will, of course), AND results in us hitting other
breakpoints, we have to single step for those ones too.  So, we have to
maintain a stack of single step operations in flight.

NOTE: this is still not quite perfect.  Suppose that process 1 gets
interrupted while in the kernel, while single stepping as described
above; and suppose that process 2 also gets interrupted in a single step
before process 1 is scheduled again.  Suppose further that process 2
gets scheduled again first, and while it is still in the kernel it
encounters the breakpoint that process 1's kernel thread was single
stepping over.  Then our little stack is going to get messed up.

So, technically we need to make this stack pid-indexed.  But not right
now :).