Implement a more precise syscall return probe for a couple of syscalls.
For sys_socketcall and sys_waitpid, we put the return check on the single
"ret" instructions in both. For our use in A3, this allows us to NOT use
the generic syscall return probe which will be triggered by every syscall
made. Aaron reports a 20x speed up in the guest for our A3 configuration.

Also made a couple of global data structures (outstanding syscalls and
current processes) per-domain, since that is what they should be. Won't
matter in our use right now since we only monitor one domain.