[CTXPROBES] Provide user probe handlers with return address in stack
User have now three different types of probes to register:

    1) call probe
    2) prologue probe
    3) return probe

The framework grabs the return address from the top of the stack and
passes it to both user's prologue and return probe handlers, with an
exception of call probe handler.

NOTE: I found that SP value was NOT the same when the function proloque
was invoked and when the function returned, though it was supposed to
be the same as it was tested in target/dumptarget.  Since I could not
figure out why it was happening, I temporarily save the SP value read
upon the prologue invocation in a global variable, and simply re-use
the saved SP to get return address upon the return of the function.