Add type compression and debugfile load option handling.
TYPE COMPRESSION:

Add support for shared type symbols, to enable DWARF "type compression".
Now, each debugfile struct has a global "shared types" symtab that any
named (non-enum) type from one compilation unit can be promoted to, and
then reused by other compilation units.  Since CUs are designed to be
fully independent of each other, DWARF debug_info sections can contain a
lot of type symbol duplication.  A lot.  So, if we can match a newly
parsed type with one that has been parsed before, we can save lots of
memory (and potentially CPU time during debugfile parsing).

However, this is rather complicated and required a bit of re-jiggering
of the library... since at least this part of its design (independent,
per-CU symtabs) was a direct map to DWARF.  Change the symtab of a
symbol (especially after it is already assigned and hashed into our
structs) is difficult.  It's a bit easier since we make the assumption,
for now, that only type symbols can be shared.

The main side effect of this is that dwdebug now holds refs to shared
symbols internally so it can tell when it is safe to free them.  This
should not matter to library users, but it does complicate the library.
I had hoped to avoid refcnt usage inside the library, but it basically
can't be done if we want to share symbols to reduce memory duplication.

You can enable type compression with a couple different flags:
the single flag REDUCETYPES, or the tuple
REDUCETYPES,REDUCETYPES_FULL_EQUIV.  The single-flag option enables very
fast, name-based type equivalence.  For instance, 'struct foo' from
bar.c is equivalent to 'struct foo' from baz.c.  This is probably fairly
safe in many cases, but obviously it will not be safe for all!  However
-- it saves both memory allocation and CPU processing time, since name
equivalence can be trivially checked during parsing, and if one name
matches another previously-shared name, we can skip to the next DIE
after the type -- and if that type was a struct type with members, we
have saved a whole bunch of memory allocation, parsing, etc.

But, name equivalence is a lie and obviously may not work for all C
programs, so we also support full type equivalence.  The problem with
full type equivalence is that 1) we have to fully parse and load each
type before checking equivalence, and 2) it requires multiple passes
over all loaded symbols following the main parsing loop.  Those passes
are faster than the main loop, of course, but still.

Obviously, type compression required other things, including exposing
the setting of debugfile load options.

DEBUGFILE LOAD OPTIONS:

rfilters have had their struct names reversed to make more sense; the
'struct rfilter' is the primary thing (no more 'struct rfilter_list');
and the 'struct rfilter_entry' takes the place of 'struct rfilter'.
There is some new code for easily creating rfilters from strings (of
course the rfilter_entry items won't have a (void *) data rock, but that
doesn't really need to be there for some of the filters we want.

Since we are enabling more debugfile load options, there is now code to
parse and free debugfile load option strings and free them; you can pass
them into `dumpdebuginfo' and `dumptarget' with the -F argument (and you
can specify as many -F arguments as you want; they chain together).  The
format of these strings is

  FLAG,FLAG,FLAG|| \
  [<A|1|R|0>::][<A|1|R|0>:]dfn;[<A|1|R|0>:]dfn;[<A|1|R|0>:]dfn|| \
  [<A|1|R|0>::][<A|1|R|0>:]sfn;[<A|1|R|0>:]sfn;[<A|1|R|0>:]sfn|| \
  [<A|1|R|0>::][<A|1|R|0>:]ss;[<A|1|R|0>:]ss;[<A|1|R|0>:]ss

where the `FLAG's are from dwdebug/dwdebug.h, and the other three
subparts of the string (separated by ||'s) specify rfilters for
debuginfo filenames; source code filenames (i.e., each DWARF CU has a
different source filename); and symbol strings.

The idea with an rfilter is that you specify an optional A or 1
(accept), or R or 0 (reject) to specify the default behavior of the
filter -- accept (i.e., apply this filter and its options to the object
we're trying to match it against), or reject (i.e., do NOT apply this
filter and its options to the object we're trying to match it against).
The default policy for the filter is used if none of the filter entries
(the items separated by `;' characters) match the string being checked.
Right, filter entries.  Each entry can also have an accept/reject
action, and the entry itself is a POSIX regular expression.

Example: suppose you give two -F args to dumptarget when loading an
executable with path /tmp/dummy that links with libc:

  -F 'REDUCETYPES||R::A:dummy' -F 'PARTIALSYM,REDUCETYPES||R::A:libc'

then when the debuginfo files are loaded for dummy, type compression
will be enabled (but only named-based equivalence); when they are loaded
for libc, partial symbol loading will occur, and type compression will
be enabled.  The first filter will not be used for loading libc's
debuginfo file, and the second filter will not be used for loading
dummy's debuginfo file.  This is a little weird, I know... but I think
it's flexible.

There are some unrelated things, but not a ton, relatively speaking.