[CTRL-FLOW] Catch first task switch to every suspected process
Catch the first task switch to every suspected process of the given
PID list to put a watchpoint at the UID of each of them, only if the
UID is non-root.  This strategy doesn't seem to work.  The UID of
each of the suspected processes turns out to be zero (root) at the
first task switch to the process.  We should then detect the creations
of the suspectecd processes and put watchpoints.