[CTRL-FLOW] Go back to instrumenting first context switches from fork
Mike and Anton found my stupid mistake to run linux-sendpage as root
in the guest VM.  Update the Makefile with branch counter and PIDs from
newly recorded replay session.  Go back to the code that instruments
the first context switch to each of the suspected processes to put an
watchpoint at uid.