Make partially loaded debugfiles and symbols expandable.
Right now, expansion is invisible and automatic for the user if they use
the debugfile_lookup_addr and debugfile_lookup_sym functions; if they
lookup a partially-loaded symbol (or a member of a partially-loaded
symbol), all the symbols in the lookup path will become fully loaded if
they match.  If the user was looking up an address, and the CU the
address is in was not fully loaded, it will be fully loaded (we have to
load every function in order to provide the tightest bound symbol to the
address, which is almost certainly what is desired) (in the future, we
can optimize to only load functions and their type dependencies, but
this is likely not to save us much memory, so it's probably not worth
doing -- there are not a lot of globals in most programs, and the
globals and their dependent symbols are the only things that might be
saved).

Basically, if a user loads a debugfile partially via
DEBUGFILE_LOAD_FLAG_CUHEADERS or DEBUGFILE_LOAD_FLAG_PUBNAMES, or loads
symbols partially via DEBUGFILE_LOAD_FLAG_PARTIALSYM (or a combination
of those), we keep the ELF debuginfo file and DWARF descriptors around
forever; then we can incrementally load the file.

If all you care about is doing address lookups, the best way to proceed
is to initially load the file with DEBUGFILE_LOAD_FLAG_CUHEADERS; then
when you lookup addresses, only those CUs that have the addresses you
care about will be fully loaded (and we can of course optimize this
more, when addresses and symbol names are all we care about -- but that
gets dangerous in that symbol types would then no longer be present --
so the user would have to be careful and we might choose not to do this
optimization... it just would make address lookups as fast as possible).

If you care about doing symbol lookups, and you know your symbols (or at
least the top-level symbol name, if looking up nested symbols) are in
the .debug_pubnames section of the debuginfo file (which you can check
via `readelf --debug-dump=pubnames'), you should load with options
DEBUGFILE_LOAD_FLAG_PUBNAMES and DEBUGFILE_LOAD_FLAG_PARTIALSYM; then
the details of your symbol will be loaded via symbol autoexpansion
during lookup; this will provide huge speed and memory savings.  If you
don't know if your symbols are all in .debug_pubnames, you can still use
DEBUGFILE_LOAD_FLAG_PARTIALSYM safely IF you always lookup symbols from
a top-level symbol -- i.e., one that is accessible from the source
file's "global" scope.  You won't get as much savings... but you can
also use CU filters to deliberately not load some CUs you know you don't
care about (i.e., if you know you'll never care about ICMP functions in
the kernel's network stack).

debugfile_expand_cu and debugfile_expand_symbol do the work of expanding
CUs fully or in part.  Users should not call them directly; users should
only see fully-loaded symbols when they query the library.  But I may
not have done the autoexpansion for all the query path functions yet.