"Dynamic" symbols are now "synthetic" symbols.
target_create_dynamic_type_pointer is now target_create_synthetic_type_pointer.

Again, synthetic symbols are user-created symbols that didn't necessarily exist
in the debuginfo file.  Users create them for convenience, i.e., if they have to
load a raw address that is a pointer to a type in the debuginfo, or something.