[CTXPROBES] Put watchpoints on raw addresses of schedule.next->uid.
The only way to put watchpoints at schedule.next->uid is using the raw
addresses at the moment.  Also, we can only use HW watchpoints that
are limited to four -- this is OK for now since we will only need one
watchpoint for one of the suspected chain of processes, which has
non-root uid.