Bug: only compute base_addr for function/label loclists.
When parsing dwarf location list attrs for DIE symbols to compute
the symbol's minimum address, I was not checking to see if the symbol
was a function or label -- only those symbols need a base address
computed from their loclists (i.e., if the function bytes are spread
over multiple text segments, it'll be represented as a list).  But,
var loclists also have a dwarf location expression associated with the
start/end addrs in the loclist; the loc expr indicates how to find the
var's real address given that EIP is in one of the start/end tuples in
the loclist.

So we were entering vars into the debugfile->addresses optimization
hashtable mistakenly, because they were getting base_addr values when
they should not have!