Add target_addressof_bsymbol.
target_addressof_bsymbol is the first thing that can operate on
pointer-indirected bsymbol chains.  So, if you have

  struct foo { int x; int **y; };
  struct bar { struct foo *foos; };
  struct bar bar_instance;

debugfile_lookup_sym("bar_instance.foos.x") will produce an lsymbol
chain that can be used to autoload across the pointer.  Of course,
bsymbol_load can't do that, but target_addressof_bsymbol can!  These
kinds of lsymbol chains will cause problems for many of our functions,
but they'll be updated like target_addressof_bsymbol eventually.

target_addressof_bsymbol traverses through the lsymbol, loading nested
chunks.  If the end of the chain is a function, we return its base
address.  Otherwise, we do nothing for functions (they may be used to
resolve the location of variables, however -- i.e., computing the dwarf
frame_base virtual register).  Otherwise, for variables, if their types
are pointers, we load the pointer, and keep loading the chain according
to the next type.  If the last var is a pointer, and the AUTO_DEREF flag
is set, we deref the pointer(s) and return the value of the last
pointer.  If it is not a pointer, we return the computed address of the
last var.

Also, this includes moving location_resolve_member_offset from target to
location_resolve_offset in dwdebug; it's not really a target-dependent
operation.