[CTRL-FLOW] Add pass 5 and instrument handle_pte_fault.
This fails because a segfault occurs.  More precisely, it occurs when
loading function arguments of handle_pte_fault, exactly here:

  struct bsymbol tbsym = {
      .lsymbol = &tlsym,
      .region = probe->probepoint->range->region
  };

Probably one of the pointers in 'probe->probepoint->range->region' is
NULL or something invalid.

My program prints this debug message right before the segfault:

  7 (hotplug): __INLINED(ref34b098:handle_pte_fault) called (ctxt: TRAP)

Given that, one guess is that the problem is releated to inlined
functions.  Dave, have you loaded arguments of a inlined function using
target?  Is target capable of doing that?

Another guess is that it is releated with the fact that the probe on
handle_pte_fault gets triggered when the context is TRAP.  Context
switches to TRAP when do_page_fault gets called just before
handle_pte_fault is called.  Well, but it is unlikely that this is the
cause of the problem since the segfault occurs while loading function
arguments.