[CTXPROBES] Use trigger->probepoint if probe->probepoint is NULL.
Use trigger->probepoint if probe->pointpoint is NULL when loading
function arguments because the probe is a high-level probe.  This
does fix the problem that we couldn't load the arguments of the
inlined function handle_pte_fault.

Will try loading the local variables of do_page_fault, passed to
handle_pte_fault instead for the time being while Dave is working to
fix the problem.